# fleet-api/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Mock Fleet API
  version: 1.5.1
  description: REST contract for mock lifecycle, configuration, and persisted mapping files.
paths:
...
  /__fleet/api/mocks/{mockId}/start:
    parameters:
      - $ref: "#/components/parameters/MockId"
    post:
      operationId: startMock
      summary: Idempotently start a mock
      responses:
        "200":
          description: The mock is already RUNNING.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockLifecycleResponse"
        "202":
          description: The mock is STARTING. Poll this operation after retryAfterMs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockLifecycleResponse"
        "400":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
        "503":
          $ref: "#/components/responses/ApiError"
  /__fleet/api/mocks/{mockId}:
    parameters:
      - $ref: "#/components/parameters/MockId"
    delete:
      operationId: stopMock
      summary: Idempotently stop a RUNNING, STARTING, FAILED, or absent mock
      responses:
        "200":
          description: The mock is STOPPED. Repeated calls return the same state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockLifecycleResponse"
        "400":
          $ref: "#/components/responses/ApiError"
        "500":
          $ref: "#/components/responses/ApiError"
...
components:
  parameters:
    MockId:
      name: mockId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
  responses:
    ApiError:
      description: Structured API failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    MockLifecycle:
      type: string
      enum: [STOPPED, STARTING, RUNNING, FAILED]
...
    MockLifecycleResponse:
      type: object
      required: [mockId, status, podName, message, retryAfterMs]
      properties:
        mockId: { type: string }
        status:
          $ref: "#/components/schemas/MockLifecycle"
        podName: { type: [string, "null"] }
        message: { type: [string, "null"] }
        retryAfterMs:
          type: [integer, "null"]
          minimum: 0
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, retryable, stateMayHaveChanged, details]
      properties:
        code: { type: string }
        message: { type: string }
        retryable: { type: boolean }
        stateMayHaveChanged: { type: boolean }
        details:
          type: object
          additionalProperties: true
...